+++ src/Views/Shared/AddressManual.cshtml
@@ -1,4 +1,4 @@
-﻿@using form_builder.Models.Elements;
+﻿@using form_builder.Constants;
 @model form_builder.Models.Elements.AddressManual;
 @{    
     var inputClass = "govuk-input";
